crypto_bot04/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiohttp==3.9.1
pydantic==2.5.0
PyYAML==6.0.1
python-dotenv==1.0.0

# CORS support for frontend
fastapi-cors==0.0.6

# Trading and crypto dependencies
ccxt==4.1.77
coinbase-advanced-py==1.8.2
openai==1.12.0

# Data handling
pandas==2.1.4
numpy==1.24.4

# Async support
asyncio-mqtt==0.11.1
async-timeout==4.0.3 